auth the user for edit the project

+++ app/Http/Controllers/ProjectController.php
@@ -61,6 +61,9 @@ class ProjectController extends Controller
      */
     public function show(Project $project)
     {
+        if(auth()->user()->id !== $project->user_id){
+            abort(403);
+        }
         return view('projects.show',compact('project'));
     }
 
@@ -72,6 +75,9 @@ class ProjectController extends Controller
      */
     public function edit(Project $project)
     {
+        if(auth()->user()->id !== $project->user_id){
+            abort(403);
+        }
         return view('projects.edit',compact('project'));
     }
 
@@ -104,6 +110,10 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
+        if(auth()->user()->id !== $project->user_id){
+            abort(403);
+        }
+        
         $project->delete();
         return redirect()->route('projects.index');
     }
